+++ ballcat-common/ballcat-common-core/src/main/java/com/hccake/ballcat/common/core/jackson/NullSerializerModifier.java
@@ -15,26 +15,43 @@ import java.util.Map;
  * @version 1.0
  * @date 2019/10/17 23:19
  */
-public class ArraySerializerModifier extends BeanSerializerModifier {
+public class NullSerializerModifier extends BeanSerializerModifier {
 
 	private final JsonSerializer<Object> nullArrayJsonSerializer = new NullArrayJsonSerializer();
 
 	private final JsonSerializer<Object> nullMapJsonSerializer = new NullMapJsonSerializer();
+
+	private final JsonSerializer<Object> nullStringJsonSerializer = new NullStringJsonSerializer();
 
 	@Override
 	public List<BeanPropertyWriter> changeProperties(SerializationConfig config, BeanDescription beanDesc,
 			List<BeanPropertyWriter> beanProperties) {
 		// 循环所有的beanPropertyWriter
 		for (BeanPropertyWriter writer : beanProperties) {
-			// 判断字段的类型，如果是array，list，set则注册nullSerializer
-			if (isArrayType(writer)) {
+			if (isStringType(writer)) {
+				// null 字符串转空
+				writer.assignNullSerializer(this.nullStringJsonSerializer);
+			}
+			else if (isArrayType(writer)) {
+				// null array 或 list，set则注册nullSerializer
 				writer.assignNullSerializer(this.nullArrayJsonSerializer);
 			}
 			else if (isMapType(writer)) {
+				// null Map 转 {}
 				writer.assignNullSerializer(this.nullMapJsonSerializer);
 			}
 		}
 		return beanProperties;
+	}
+
+	/**
+	 * 是否是 String 类型
+	 * @param writer BeanPropertyWriter
+	 * @return boolean
+	 */
+	private boolean isStringType(BeanPropertyWriter writer) {
+		Class<?> clazz = writer.getType().getRawClass();
+		return String.class.isAssignableFrom(clazz);
 	}
 
 	/**