+++ ballcat-common/ballcat-common-core/src/main/java/com/hccake/ballcat/common/core/jackson/NullStringJsonSerializer.java
@@ -11,7 +11,7 @@ import java.io.IOException;
  * @version 1.0
  * @date 2019/10/17 22:19 jackson NULL值序列化为 ""
  */
-public class NullSerializer extends JsonSerializer<Object> {
+public class NullStringJsonSerializer extends JsonSerializer<Object> {
 
 	@Override
 	public void serialize(Object value, JsonGenerator jsonGenerator, SerializerProvider provider) throws IOException {